+++ src/hexsample/source.py
@@ -183,20 +183,63 @@ class PointBeam(AbstractBeam):
         x = np.full(size, self.x0)
         y = np.full(size, self.y0)
         return x, y
-    
 
-class FlatBeam(HexagonalGrid, AbstractBeam):
-    
-    """Flat beam that illuminates uniformly the entire chip.
+
+@dataclass
+class SquareBeam(AbstractBeam):
+
+    """Square uniform X-ray beam.
+
+    Arguments
+    ---------
+    x0 : float
+        The x-coordinate of the beam centroid in cm.
+
+    y0 : float
+        The y-coordinate of the beam centroid in cm.
+
+    side : float
+        The square width in cm.
     """
+
+    side: float = 1.52
 
     def rvs(self, size: int = 1) -> Tuple[np.ndarray, np.ndarray]:
         """Overloaded method.
         """
-        x_min, y_max = self.pixel_to_world(0, 0)
-        x_max, y_min = self.pixel_to_world(self.num_cols - 1, self.num_rows - 1)
-        x = rng.generator.uniform(x_min, x_max, size=size)
-        y = rng.generator.uniform(y_min, y_max, size=size)
+        x = rng.generator.uniform(self.x0 - self.side/2., self.x0 + self.side/2., size=size)
+        y = rng.generator.uniform(self.y0 - self.side/2., self.y0 + self.side/2., size=size)
+        return x, y
+
+
+@dataclass
+class RectangleBeam(AbstractBeam):
+
+    """Rectangle uniform X-ray beam.
+
+    Arguments
+    ---------
+    x0 : float
+        The x-coordinate of the beam centroid in cm.
+
+    y0 : float
+        The y-coordinate of the beam centroid in cm.
+    
+    width : float
+        The rectangle width in cm.
+    
+    height : float
+        The rectangle height in cm.
+    """
+
+    width: float = 1.52
+    height: float = 1.52
+
+    def rvs(self, size: int = 1) -> Tuple[np.ndarray, np.ndarray]:
+        """Overloaded method.
+        """
+        x = rng.generator.uniform(self.x0 - self.width/2., self.x0 + self.width/2., size=size)
+        y = rng.generator.uniform(self.y0 - self.height/2., self.y0 + self.height/2., size=size)
         return x, y
 
 
@@ -399,7 +442,8 @@ BeamProxy.register("gaussian", GaussianBeam, default=True)
 BeamProxy.register("slit", SlitBeam)
 BeamProxy.register("triangular", TriangularBeam)
 BeamProxy.register("hexagonal", HexagonalBeam)
-BeamProxy.register("flat", FlatBeam)
+BeamProxy.register("square", SquareBeam)
+BeamProxy.register("rectangle", RectangleBeam)
 
 
 @dataclass